refactor(pwa): single ESM entry, address review findings
- Shell loads from one <script type="module"> instead of five separate tags
- Cold offline boot works: service worker caches the versioned paths the page actually requests
- Mobile shell stops re-scanning the DOM on every dungeon redraw
- ?pwa=0 disables the shell and PWA assets no longer ship to non-PWA visitors
- Pinch-zoom and assistive-tech gestures preserved on the dungeon viewport
- Hold-to-repeat survives sliding off the edge of the touch keyboard
- Minimap defers to PWA mode at one early-return instead of nine callsites
- Status-strip folded back into controls (94 lines, not worth its own seam)
