+++ crawl-ref/source/webserver/templates/client.html
@@ -8,9 +8,11 @@
     <meta name="viewport" content="width=860, initial-scale=1.0">
 {% end %}
     <link rel="icon" href="/static/stone_soup_icon-32x32.png" type="image/png">
+{% if pwa_enabled %}
     <link rel="manifest" href="/static/pwa/manifest.webmanifest">
     <link rel="apple-touch-icon" href="{{ static_url("pwa/icons/apple-touch-icon.png") }}">
     <meta name="theme-color" content="#000000">
+{% end %}
     <script type="text/javascript">
       {% set fail_safe_game_version = globals().get('game_version', '') %}
       var socket_server = "{{ socket_server }}";
@@ -19,15 +21,13 @@
     <script src="{{ static_url("scripts/contrib/require.js") }}" data-main="/static/scripts/app"></script>
     <link rel="stylesheet" type="text/css" href="{{ static_url("style.css") }}">
     <link rel="stylesheet" type="text/css" href="{{ static_url("fonts/fonts.css") }}">
+{% if pwa_enabled %}
     <link rel="stylesheet" type="text/css" href="{{ static_url("pwa/mobile-pwa.css") }}">
-    <script src="{{ static_url("pwa/mobile-pwa-control-keymaps.js") }}"></script>
-    <script src="{{ static_url("pwa/mobile-pwa-status-strip.js") }}"></script>
-    <script src="{{ static_url("pwa/mobile-pwa-controls.js") }}"></script>
-    <script src="{{ static_url("pwa/mobile-pwa-dungeon-touch.js") }}"></script>
-    <script src="{{ static_url("pwa/mobile-pwa.js") }}"></script>
+    <script type="module" src="{{ static_url("pwa/mobile-pwa.js") }}"></script>
     <meta name="apple-mobile-web-app-capable" content="yes">
     <meta name="mobile-web-app-capable" content="yes">
     <meta name="apple-mobile-web-app-status-bar-style" content="black-translucent">
+{% end %}
   </head>
   <body>
     <noscript>Please enable javascript!</noscript>
